build-requirements.txt
pyinstaller>=5.13.0
pillow>=10.0.0
cairosvg>=2.7.1  # For icon conversion if needed